+++ packages/jaspr_content/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased patch
 
+- Output the title and description from page data in `MarkdownOutput`.
 - Update `package:mustache_template` constraint to `^2.0.2`.
   This version adds support for emojis in templates,
   no longer requiring a dependency override.